ESK-001 Updated to latest dependencies, changed config module, added favicon
