+++ .eslintrc
@@ -1,38 +1,17 @@
 {
-  rules: {
-    indent: [2, 2, {"SwitchCase": 1}],
-    no-multi-spaces: 2,
-    no-underscore-dangle: 0,
-    no-use-before-define: [1, "nofunc"],
-    no-unused-expressions: 0,
-    no-empty-class: 0,
-    object-curly-spacing: [2, "always"],
-    quotes: [1, "single"],
-    space-in-parens: [2, "never"]
-  },
-  env: {
-    node: true,
-    es6: true
-  },
-  globals: {
-    angular: true,
-    $: true,
-    jQuery: true,
-    moment: true,
-    window: true,
-    document: true,
-    Modernizr: true,
-    __TESTING__: true,
-    beforeEach: true,
-    expect: true,
-    describe: true,
-    it: true,
-    element: true,
-    by: true,
-    browser: true,
-    inject: true,
-    register: true,
-    sinon: true,
-    _: false
+  "extends": "airbnb-base",
+  "rules": {
+    "comma-dangle": ["error", "never"],
+    "class-methods-use-this": "off",
+    "object-curly-newline": [
+      "error",
+      {
+        "consistent": true,
+        "multiline": true
+      }
+    ],
+    "radix": ["error", "as-needed"],
+    "no-param-reassign": ["error", { "props": false }],
+    "max-len": ["error", { "code": 100, "ignoreComments": true }]
   }
 }
