--- api/conf.js
@@ -1,9 +1,0 @@
-/**
- * Created by Andriy Ermolenko on 17.08.15.
- */
-
-const nconf = require('nconf');
-
-nconf.argv().env().file({file: './config.json' });
-
-module.exports = nconf;
