+++ config/default.json
@@ -1,0 +1,5 @@
+{
+  "app": {
+    "port": "3000"
+  }
+}
