+++ package.json
@@ -1,16 +1,28 @@
 {
   "name": "express-starter-kit",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "dependencies": {
     "body-parser": "^1.17.2",
+    "bower": "^1.8.8",
+    "config": "^3.2.2",
     "ejs": "^2.5.6",
     "express": "^4.15.3",
-    "method-override": "^2.3.5",
-    "nconf": "^0.8.4"
+    "method-override": "^3.0.0",
+    "serve-favicon": "^2.5.0"
+  },
+  "devDependencies": {
+    "eslint": "^6.0.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.18.0"
   },
   "scripts": {
-    "start": "node app.js",
+    "start": "node index",
+    "lint": "eslint src",
     "postinstall": "bower install --config.interactive=false"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/invercity/express-starter-kit.git"
   },
   "license": "MIT"
 }
