--- routes/index.js
@@ -1,7 +1,0 @@
-/**
- * Created by Andriy Ermolenko on 17.08.15.
- */
-
-module.exports = {
-  index: (req, res) => res.render('index'),
-};
