+++ src/routes/index.js
@@ -1,0 +1,7 @@
+/**
+ * Created by Andrii Yermolenko on 17.08.15.
+ */
+
+module.exports = {
+  index: (req, res) => res.render('index')
+};
